requirements.txt
finnhub-python
pandas
google-genai
python-dotenv
lxml
tabulate